+++ packages/api-apw/src/plugins/utils.ts
@@ -1,5 +1,4 @@
 import get from "lodash/get";
-import { customAlphabet } from "nanoid";
 import { CmsModelField } from "@webiny/api-headless-cms/types";
 import { SecurityIdentity } from "@webiny/api-security/types";
 import {
@@ -15,9 +14,6 @@ import {
     ApwWorkflowStepTypes,
     WorkflowScopeTypes
 } from "~/types";
-
-const ALPHANUMERIC = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-export const getNanoid = customAlphabet(ALPHANUMERIC, 10);
 
 export interface CreateModelFieldParams
     extends Omit<CmsModelField, "id" | "storageId" | "fieldId"> {